requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.8.1
django-htmx==1.13.0
django-js-asset==2.0.0
django-timezone-field==5.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
kombu==5.3.1
openpyxl==3.1.2
outcome==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
redis==4.4.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.2
trio-websocket==0.10.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
djangorestframework==3.14.0
